requirements.txt
torch==2.3.0
-f https://download.pytorch.org/whl/cpu
langgraph
langchain
langchain-community
python-dotenv
fastapi
motor
langchain-google-genai
langchain-openai
langchain-pinecone
PyMuPDF
pytz
loguru
uvicorn[standard]
python-multipart
langchain_experimental
duckduckgo-search
pydantic[email]
python_jose==3.3.0
pillow
python-docx
langchain-text-splitters
python-docx
cloudinary
PyJWT==2.8.0
langchain_docling